Define Artemis grammer
We define the grammar of Artemis to be a simple one: it can be used
to print one or more arithmetic expressions using it.

This simple grammar will make it easier to develop the whole
compiler stack, without having to worry about implementing support
for variables/functions and various other common language features.